project-backend-master/requirements.txt
appdirs==1.4.4
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
distlib==0.3.4
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
idna==3.3
imgspy==0.2.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==2.1.1
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
requests==2.27.1
simplejson==3.16.0
six==1.16.0
urllib3==1.26.9
virtualenv==20.4.2
Werkzeug==2.1.1
zipp==3.8.0
